+++ app/dashboards/preregistration_dashboard.rb
@@ -1,0 +1,87 @@
+# frozen_string_literal: true
+
+require 'administrate/base_dashboard'
+
+# Administrate dashboard for Preregistrations
+class PreregistrationDashboard < Administrate::BaseDashboard
+  # ATTRIBUTE_TYPES
+  # a hash that describes the type of each of the model's fields.
+  #
+  # Each different type represents an Administrate::Field object,
+  # which determines how the attribute is displayed
+  # on pages throughout the dashboard.
+  ATTRIBUTE_TYPES = {
+    event: Field::BelongsTo,
+    id: Field::Number,
+    first_name: Field::String,
+    last_name: Field::String,
+    email: Field::String,
+    net_id: Field::String,
+    upi: Field::String,
+    phone: Field::String,
+    checked_in: Field::Boolean,
+    created_at: Field::DateTime,
+    updated_at: Field::DateTime
+  }.freeze
+
+  # COLLECTION_ATTRIBUTES
+  # an array of attributes that will be displayed on the model's index page.
+  #
+  # By default, it's limited to four items to reduce clutter on index pages.
+  # Feel free to add, remove, or rearrange items.
+  COLLECTION_ATTRIBUTES = %i(
+    event
+    net_id
+    first_name
+    last_name
+  ).freeze
+
+  # SHOW_PAGE_ATTRIBUTES
+  # an array of attributes that will be displayed on the model's show page.
+  SHOW_PAGE_ATTRIBUTES = %i(
+    event
+    id
+    first_name
+    last_name
+    email
+    net_id
+    upi
+    phone
+    checked_in
+    created_at
+    updated_at
+  ).freeze
+
+  # FORM_ATTRIBUTES
+  # an array of attributes that will be displayed
+  # on the model's form (`new` and `edit`) pages.
+  FORM_ATTRIBUTES = %i(
+    event
+    first_name
+    last_name
+    email
+    net_id
+    upi
+    phone
+    checked_in
+  ).freeze
+
+  # COLLECTION_FILTERS
+  # a hash that defines filters that can be used while searching via the search
+  # field of the dashboard.
+  #
+  # For example to add an option to search for open resources by typing "open:"
+  # in the search field:
+  #
+  #   COLLECTION_FILTERS = {
+  #     open: ->(resources) { where(open: true) }
+  #   }.freeze
+  COLLECTION_FILTERS = {}.freeze
+
+  # Overwrite this method to customize how preregistrations are displayed
+  # across all pages of the admin dashboard.
+  #
+  # def display_resource(preregistration)
+  #   "Preregistration ##{preregistration.id}"
+  # end
+end